requirements.txt
fastapi
uvicorn
pymupdf
serpapi
pydantic
python-dotenv
sqlalchemy
psycopg2-binary
pandas
numpy
scikit-learn
python-multipart
spacy
sentence-transformers
faiss-cpu
xgboost
apscheduler
transformers
huggingface-hub
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
